Factory cleanup fixes (#3312)
* use --parent-death when launching workers from vine_factory

* ignore error on factory cleanup, fallback to atexit function

* factory to use different staging dir as manager

This is because the manager may clean up before the factory ends.

* remove final / in runtime dirs

Otherwise os.path.dirname gets confused. We should we using
pathlib.Path anyway...

* simplify atexit cleanup

* plug memory leak

+++ taskvine/src/manager/vine_runtime_dir.c
@@ -110,6 +110,12 @@ char *vine_runtime_directory_create() {
     register_staging_dir(tmp);
 	free(tmp);
 
+	tmp = string_format("%s/../vine-cache", runtime_dir);
+	if(!create_dir(tmp, 0755)) {
+        return NULL;
+    }
+	free(tmp);
+
     if(symlink_most_recent) {
         char *tmp = path_concat(vine_runtime_info_path, "most-recent");
         unlink(tmp);
@@ -121,16 +127,25 @@ char *vine_runtime_directory_create() {
 }
 
 char *vine_get_runtime_path_log(struct vine_manager *m, const char *path) {
-    return string_format("%s/vine-logs/%s", m->runtime_directory, path ? path : "");
+    return string_format("%s/vine-logs%s%s",
+			m->runtime_directory,
+			path ? "/" : "",
+			path ? path : "");
 }
 
 char *vine_get_runtime_path_staging(struct vine_manager *m, const char *path) {
-    return string_format("%s/staging/%s", m->runtime_directory, path ? path : "");
+    return string_format("%s/staging%s%s",
+			m->runtime_directory,
+			path ? "/" : "",
+			path ? path : "");
 }
 
 char *vine_get_runtime_path_caching(struct vine_manager *m, const char *path) {
     char abs[PATH_MAX];
-    char *tmp = string_format("%s/../vine-cache/%s", m->runtime_directory, path ? path : "");
+    char *tmp = string_format("%s/../vine-cache%s%s",
+			m->runtime_directory,
+			path ? "/" : "",
+			path ? path : "");
     path_collapse(tmp, abs, 1);
     free(tmp);
     return xxstrdup(abs);
